buarque_agent_langgraph/pyproject.toml
[project]
name = "a2a-buarque-agent-langgraph"
version = "0.1.0"
description = "Buarque's LangGraph scheduling agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.5,<0.3.0",
    "httpx>=0.28.1",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain-core",
    "langchain-openai>=0.3.33",
    "openai>=1.108.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
